+++ src/components/Hero/heroChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: [
+      'Updates the validation of the background props in the "campaign" variant.',
+    ],
+  },
+  {
     date: "2025-09-30",
     version: "4.0.2",
     type: "Update",