+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Updates the `TagSet` component to add semantic color variants.
 - Updates the hex color value for `ui.warning.tertiary` for better accessibility compliance.
+- Updates the `Hero` component to alter the validation of the background props in the `"campaign"` variant.
 
 ### Fixes